Layout overhaul
Latest Emacs emits a warning when load-path contains the ~/.emacs.d directory
(see http://lists.gnu.org/archive/html/emacs-devel/2012-03/msg00056.html for
details).

So I took the opportunity to revisit the layout of files, moving most
sources from the root directory to the esk/ subdirectory, and loading
them directly instead of require-ing them, to avoid the need to add
the directory to the load-path.